Corrected uncertainty calculations, tweaks for less hard-coding in dataBank(), corrected regressionScale() method, changed all inputs to transitModel.occultquad() to c_doubles, simulated data's now reproduces the injected light curve to within 1-sigma when using prayer-bead uncertainty estimation

+++ Code/oscaar/c/transit1forLMLS.c
@@ -54,7 +54,8 @@
 
 #define pi 3.14159265358979311600
 
-void occultquad(double *t, float p, float ap, float P, float i, float gamma1, float gamma2, double e, double longPericenter, double t0, float n, double *F);
+//void occultquad(double *t, float p, float ap, float P, float i, float gamma1, float gamma2, double e, double longPericenter, double t0, float n, double *F);
+void occultquad(double *t, double p, double ap, double P, double i, double gamma1, double gamma2, double e, double longPericenter, double t0, double n, double *F);
 
 float *linspace(float beginningPhase, float endPhase, int Npoints)
 // Replicate the function "linspace" from MATLAB
@@ -370,7 +371,9 @@ double occultuni(double z, double w)
 }
 
 
-void occultquad(double *t, float p, float ap, float P, float i, float gamma1, float gamma2, double e, double longPericenter, double t0, float n, double *F)
+//void occultquad(double *t, float p, float ap, float P, float i, float gamma1, float gamma2, double e, double longPericenter, double t0, float n, double *F)
+
+void occultquad(double *t, double p, double ap, double P, double i, double gamma1, double gamma2, double e, double longPericenter, double t0, double n, double *F)
 {
 	double t0overP;
 	double *Z, *phi; int ii;
